requirements.txt
Flask
flask_socketio
rx
kafka-python
cassandra-driver
zstandard